refactor: use context API to provide decks
